+++ CHANGELOG.md
@@ -1,10 +1,22 @@
 # Changelog
 
-## Unreleased
+## [Unreleased]
+
+### BREAKING CHANGES
+
+- Renamed all attributes and properties to one of the following format:
+  * `option-<option key>` / `option<Option Key>: For options of an constructor
+    function, e.g. `option-type` / `optionType`, `option-localematcher` /
+    `optionLocaleMatcher`, as a result, `intl-style` is now `option-style`
+  * `<method name>-<argument name>` / `<method name><Argument Name>`: For
+    arguments of a method, e.g. `format-unit` / `formatUnit`, `of-code` /
+    `ofCode`, etc.
+  * `locales`: This attribute remains unchanged
 
 ### Added
 
-- [`intl-listformat`]: Added `formatToParts()` method
+- [`intl-listformat`]: Added `formattedParts` property to return the same value
+  as `Intl.ListFormat`’s `formatToParts()` method.
 - [`intl-relativetimeformat`] Added the custom element
 - [`intl-locale`] Added the custom element